backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# RSS parsing
feedparser==6.0.10

# AI
anthropic==0.18.1

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
pydantic==2.6.1
pydantic-settings==2.1.0

# Date handling
python-dateutil==2.8.2

# HTML parsing
beautifulsoup4==4.12.3
